modified:   app/Events/ContactUsEvent.php 	modified:   app/Http/Controllers/Api/AuthController.php 	modified:   app/Http/Controllers/Api/Dashboard/BannerController.php 	modified:   app/Http/Controllers/Api/Dashboard/CategoryController.php 	modified:   app/Http/Controllers/Api/Dashboard/ProductController.php 	modified:   app/Http/Controllers/Api/Dashboard/UserController.php 	new file:   app/Http/Controllers/Api/ecommerce/ContactController.php 	modified:   app/Http/Controllers/Api/ecommerce/HomeController.php 	modified:   app/Http/Requests/ProductRequest.php 	modified:   app/Http/Resources/Dashboard/CategoryResource.php 	modified:   app/Http/Resources/Dashboard/UserDataResource.php 	modified:   app/Listeners/ContactUsListeners.php 	modified:   app/Listeners/SendMailListeners.php 	new file:   app/Mail/ContactMail.php 	new file:   app/Mail/WelcomeMail.php 	modified:   app/Models/Category.php 	modified:   app/Models/ContactUs.php 	modified:   database/migrations/2025_07_06_193236_create_categories_table.php 	modified:   database/migrations/2025_07_18_193450_create_contact_us_table.php 	deleted:    public/dashboard/assets/images/banner/68781270771bfPicture1.png 	renamed:    public/dashboard/assets/images/banner/687fffc1db3f3_Picture1.png -> public/dashboard/assets/images/banner/68814e220e2db_Picture1.png 	new file:   public/dashboard/assets/images/banner/68814e5b01bfd_Picture1.png 	new file:   public/dashboard/assets/images/banner/688299e8a0a5c_Picture1.png 	new file:   public/dashboard/assets/images/banner/688299f497623_Untitled-1 (2).jpg 	new file:   public/dashboard/assets/images/banner/68829a0260e81_Untitled-1 (2).jpg 	new file:   resources/views/emails/contact-mail.blade.php 	new file:   resources/views/emails/welcome-mail.blade.php 	modified:   routes/api.php

+++ app/Http/Controllers/Api/Dashboard/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api\Dashboard;
 use App\Helpers\ApiResponse;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Dashboard\CategoryResource;
+use App\Http\Resources\Dashboard\ProductResource;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -16,7 +17,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $data = Category::with(['parent', 'children','products'])->get();
+        $data = Category::with(['parent', 'children', 'products'])->get();
         if (!$data) {
 
             return ApiResponse::sendResponse(400, 'the category not found', []);
@@ -49,7 +50,7 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
-        //
+       
     }
 
     /**
@@ -57,7 +58,8 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
-         $validate = Validator::make($request->all(), [
+        // dd($request->all());
+        $validate = Validator::make($request->all(), [
             'category_name' => 'required|string|max:255',
             'parent_id' => 'nullable|exists:categories,id',
         ]);
@@ -76,8 +78,12 @@ class CategoryController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Category $category)
+    public function destroy($id)
     {
+        $category = Category::find($id);
+        if (!$category) {
+            return ApiResponse::sendResponse(404, 'the category not found', []);
+        }
         $deleted = $category->delete();
 
         if ($deleted) {
